refactor(detect): split chess/detect.rs into cohesive modules
detect.rs mixed three concerns. Extract the detector-agnostic NMS
helpers `is_local_max` / `count_positive_neighbors` (consumed by both
the ChESS and Radon detectors) into `detect/neighbors.rs`, and the
corner-merge logic `merge_corners_simple` / `merge_corners_naive` into
`detect/merge.rs`, with their tests. detect.rs keeps the detection
entry points and drops from ~713 to ~371 lines.

Pure move: `crate::detect::{is_local_max, count_positive_neighbors}`
and the crate-root `merge_corners_simple` re-export are unchanged.
